Create a project

Returns a sasctl MMproject object from Model Manager

create_project(
  session,
  name,
  description = NULL,
  model_function = NULL,
  input_vars = NULL,
  output_vars = NULL,
  image = NULL,
  additional_parameters = NULL,
  ...
)

Arguments

session

viya_connection object, obtained through session function

name

The name of the project

description

The description of the project.

model_function

The model function of the project. Valid values: analytical, classification, cluster, forecasting, prediction, Text categorization, Text extraction, Text sentiment, Text topics, transformation

input_vars

data.frame with the input data sample to configure the project variables.

output_vars

data.frame with the output data sample to configure the project variables.

image

Image URI to be used as project cover

additional_parameters

list of parameters and lists to be added to the payload

...

additional parameters to be passed to httr::POST such as httr::add_headers

Value

A data.frame with the list of projects

Examples

if (FALSE) {
new_project <- create_project(sess, name = "ModelProj", 
                                    description = "My fancy project",
                                    model_function = "classification")
new_project
}