Viya PUT

Wrapper to make generic DELETE calls to SAS Viya

vPUT(
  session,
  path,
  payload,
  ...,
  verbose = FALSE,
  encode = "json",
  output = "json"
)

Arguments

session

viya_connection object, obtained through session function

path

character, path to the GET api endpoint

payload

list or json string, if it is a list, will be transformed in a json string using jsonlite::toJSON

...

additional parameters to be passed to httr::PUT such as httr::add_headers

verbose

logical, return print API call information

encode

payload encoding type, to be passed to httr::POST.

output

string, if json will return httr::fromJSON(httr::content(response, as = "text")) with etag information parsed, else if response will return a httr::response() object, if text will return content(response, as = "text")

Value

list if output = "json" default. httr::response() if output = response.

Details

This function in built on top of httr for convenience when calling SAS Viya API endpoints..

See also

Examples

if (FALSE) {
folders <- vGET(session,
                path = "folders/folders/")

newFolder <- vPOST(session,
                   path = paste0("folders/folders/"),
                   query = list(parentFolderUri = folders$items$parentFolderUri[1]),
                   payload = list(name = "newFolder"),
                   httr::content_type("application/json"))

deletedFolder <- vDELETE(session,
                         path = "folders/folders/",
                         resourceID = newFolder$id)

}