Write ModelProperties json

Writes a descriptor JSON file for ModelProperties, it will configure the model properties within Model Manager

write_ModelProperties_json(
  modelName,
  modelDescription = "R model",
  modelFunction,
  trainTable = " ",
  algorithm,
  numTargetCategories,
  targetEvent,
  targetVariable,
  eventProbVar,
  modeler = " ",
  tool = "R",
  toolVersion = "default",
  path = "./",
  noFile = FALSE
)

Arguments

modelName

Name of the model

modelDescription

String describing the model

modelFunction

Classification, Prediction, Segmentation, Analytical or Clustering.

trainTable

Name of the training table

algorithm

Algorithm name (Random Forest, GLM, Linear Regression, etc.)

numTargetCategories

number of possible classes for classification

targetEvent

Target event label eg: "1", "versicolor" etc.

targetVariable

Target variable name

eventProbVar

Variable name that has the targetEvent probability Chance

modeler

Modeler's name

tool

Name of the tool used to build the model

toolVersion

Version of the tool used to build the model

path

file path where to write the json (don't include the filename)

noFile

if you don't want to write to a file, only list the output

Value

list of the mapped properties and values.

'ModelProperties.json' file written to path

Examples

write_ModelProperties_json(modelName = "My R Model", 
                           modelDescription = "Awesome Description", 
                           modelFunction = "Classification",
                           trainTable = " ",
                           algorithm = "Logistic Regression",
                           numTargetCategories = 2,
                           targetEvent = "BAD",
                           targetVariable = "P_BAD1",
                           eventProbVar = "P_BAD1",
                           modeler = "John SAS",
                           noFile = TRUE)
#>                              value
#> name                    My R Model
#> description    Awesome Description
#> function            Classification
#> scoreCodeType                    R
#> trainTable                        
#> trainCodeType                    R
#> algorithm      Logistic Regression
#> targetVariable              P_BAD1
#> targetEvent                    BAD
#> targetLevel                 Binary
#> eventProbVar                P_BAD1
#> modeler                   John SAS
#> tool                             R
#> toolVersion                  4.4.2