Write fileMetadata json

Writes a variable descriptor JSON file for fileMetadata, it will configure the models files metadata within Model Manager in the first upload

write_fileMetadata_json(
  scoreCodeName = "scoreCode.R",
  scoreResource = "model.rda",
  additionalFilesNames = c(),
  additionalFilesRoles = c(),
  path = "./",
  noFile = FALSE
)

Arguments

scoreCodeName

Name of the scoring code file

scoreResource

rda file name or other score resources.

additionalFilesNames

additional files names.

additionalFilesRoles

additional files role names.

path

filepath where to write the json (don't include the filename)

noFile

if you don't want to write to a file, only list the output

Value

list of the mapped properties and values.

'ModelProperties.json' file written to path

Examples

## Using default names and files
write_fileMetadata_json(noFile = TRUE)
#> [[1]]
#> [[1]]$role
#> [1] "inputVariables"
#> 
#> [[1]]$name
#> [1] "inputVar.json"
#> 
#> 
#> [[2]]
#> [[2]]$role
#> [1] "outputVariables"
#> 
#> [[2]]$name
#> [1] "outputVar.json"
#> 
#> 
#> [[3]]
#> [[3]]$role
#> [1] "score"
#> 
#> [[3]]$name
#> [1] "scoreCode.R"
#> 
#> 
#> [[4]]
#> [[4]]$role
#> [1] "scoreResource"
#> 
#> [[4]]$name
#> [1] "model.rda"
#> 
#> 

## addition file resources
## send 2 vectors with file names and role name, must be of same length

write_fileMetadata_json(additionalFilesNames = c("myFileName.ext", "myFileName2.ext"),
                        additionalFilesRoles = c("scoreResource", "scoreResource"),
                        noFile = TRUE
                        )
#> [[1]]
#> [[1]]$role
#> [1] "inputVariables"
#> 
#> [[1]]$name
#> [1] "inputVar.json"
#> 
#> 
#> [[2]]
#> [[2]]$role
#> [1] "outputVariables"
#> 
#> [[2]]$name
#> [1] "outputVar.json"
#> 
#> 
#> [[3]]
#> [[3]]$role
#> [1] "score"
#> 
#> [[3]]$name
#> [1] "scoreCode.R"
#> 
#> 
#> [[4]]
#> [[4]]$role
#> [1] "scoreResource"
#> 
#> [[4]]$name
#> [1] "model.rda"
#> 
#> 
#> [[5]]
#> [[5]]$role
#> [1] "scoreResource"
#> 
#> [[5]]$name
#> [1] "myFileName.ext"
#> 
#> 
#> [[6]]
#> [[6]]$role
#> [1] "scoreResource"
#> 
#> [[6]]$name
#> [1] "myFileName2.ext"
#> 
#>