Format Data.Frame rows to json format

Viya MAS requires a very specific json format which is the goal of this function to create

format_data_json(df, scr = FALSE, metadata_columns = NULL)

Arguments

df

data frame to be transformed in JSON format rows

scr

boolean, if TRUE will write the new json format with metadata

metadata_columns

columns names to be used as metadata. If scr is set to FALSE, metatada_columns is ignored

Value

a vector of JSON strings

Examples

json_output <- format_data_json(mtcars)
json_output
#> : 1
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 21 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 160 }, {\"name\": \"hp\", \"value\": 110 }, {\"name\": \"drat\", \"value\": 3.9 }, {\"name\": \"wt\", \"value\": 2.62 }, {\"name\": \"qsec\", \"value\": 16.46 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 2
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 21 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 160 }, {\"name\": \"hp\", \"value\": 110 }, {\"name\": \"drat\", \"value\": 3.9 }, {\"name\": \"wt\", \"value\": 2.875 }, {\"name\": \"qsec\", \"value\": 17.02 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 3
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 22.8 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 108 }, {\"name\": \"hp\", \"value\": 93 }, {\"name\": \"drat\", \"value\": 3.85 }, {\"name\": \"wt\", \"value\": 2.32 }, {\"name\": \"qsec\", \"value\": 18.61 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 4
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 21.4 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 258 }, {\"name\": \"hp\", \"value\": 110 }, {\"name\": \"drat\", \"value\": 3.08 }, {\"name\": \"wt\", \"value\": 3.215 }, {\"name\": \"qsec\", \"value\": 19.44 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 5
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 18.7 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 360 }, {\"name\": \"hp\", \"value\": 175 }, {\"name\": \"drat\", \"value\": 3.15 }, {\"name\": \"wt\", \"value\": 3.44 }, {\"name\": \"qsec\", \"value\": 17.02 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 6
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 18.1 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 225 }, {\"name\": \"hp\", \"value\": 105 }, {\"name\": \"drat\", \"value\": 2.76 }, {\"name\": \"wt\", \"value\": 3.46 }, {\"name\": \"qsec\", \"value\": 20.22 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 7
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 14.3 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 360 }, {\"name\": \"hp\", \"value\": 245 }, {\"name\": \"drat\", \"value\": 3.21 }, {\"name\": \"wt\", \"value\": 3.57 }, {\"name\": \"qsec\", \"value\": 15.84 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 8
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 24.4 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 146.7 }, {\"name\": \"hp\", \"value\": 62 }, {\"name\": \"drat\", \"value\": 3.69 }, {\"name\": \"wt\", \"value\": 3.19 }, {\"name\": \"qsec\", \"value\": 20 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 9
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 22.8 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 140.8 }, {\"name\": \"hp\", \"value\": 95 }, {\"name\": \"drat\", \"value\": 3.92 }, {\"name\": \"wt\", \"value\": 3.15 }, {\"name\": \"qsec\", \"value\": 22.9 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 10
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 19.2 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 167.6 }, {\"name\": \"hp\", \"value\": 123 }, {\"name\": \"drat\", \"value\": 3.92 }, {\"name\": \"wt\", \"value\": 3.44 }, {\"name\": \"qsec\", \"value\": 18.3 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 11
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 17.8 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 167.6 }, {\"name\": \"hp\", \"value\": 123 }, {\"name\": \"drat\", \"value\": 3.92 }, {\"name\": \"wt\", \"value\": 3.44 }, {\"name\": \"qsec\", \"value\": 18.9 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 12
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 16.4 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 275.8 }, {\"name\": \"hp\", \"value\": 180 }, {\"name\": \"drat\", \"value\": 3.07 }, {\"name\": \"wt\", \"value\": 4.07 }, {\"name\": \"qsec\", \"value\": 17.4 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 3 }] }"
#> ------------------------------------------------------------ 
#> : 13
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 17.3 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 275.8 }, {\"name\": \"hp\", \"value\": 180 }, {\"name\": \"drat\", \"value\": 3.07 }, {\"name\": \"wt\", \"value\": 3.73 }, {\"name\": \"qsec\", \"value\": 17.6 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 3 }] }"
#> ------------------------------------------------------------ 
#> : 14
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 15.2 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 275.8 }, {\"name\": \"hp\", \"value\": 180 }, {\"name\": \"drat\", \"value\": 3.07 }, {\"name\": \"wt\", \"value\": 3.78 }, {\"name\": \"qsec\", \"value\": 18 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 3 }] }"
#> ------------------------------------------------------------ 
#> : 15
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 10.4 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 472 }, {\"name\": \"hp\", \"value\": 205 }, {\"name\": \"drat\", \"value\": 2.93 }, {\"name\": \"wt\", \"value\": 5.25 }, {\"name\": \"qsec\", \"value\": 17.98 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 16
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 10.4 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 460 }, {\"name\": \"hp\", \"value\": 215 }, {\"name\": \"drat\", \"value\": 3 }, {\"name\": \"wt\", \"value\": 5.424 }, {\"name\": \"qsec\", \"value\": 17.82 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 17
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 14.7 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 440 }, {\"name\": \"hp\", \"value\": 230 }, {\"name\": \"drat\", \"value\": 3.23 }, {\"name\": \"wt\", \"value\": 5.345 }, {\"name\": \"qsec\", \"value\": 17.42 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 18
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 32.4 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 78.7 }, {\"name\": \"hp\", \"value\": 66 }, {\"name\": \"drat\", \"value\": 4.08 }, {\"name\": \"wt\", \"value\": 2.2 }, {\"name\": \"qsec\", \"value\": 19.47 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 19
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 30.4 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 75.7 }, {\"name\": \"hp\", \"value\": 52 }, {\"name\": \"drat\", \"value\": 4.93 }, {\"name\": \"wt\", \"value\": 1.615 }, {\"name\": \"qsec\", \"value\": 18.52 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 20
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 33.9 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 71.1 }, {\"name\": \"hp\", \"value\": 65 }, {\"name\": \"drat\", \"value\": 4.22 }, {\"name\": \"wt\", \"value\": 1.835 }, {\"name\": \"qsec\", \"value\": 19.9 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 21
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 21.5 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 120.1 }, {\"name\": \"hp\", \"value\": 97 }, {\"name\": \"drat\", \"value\": 3.7 }, {\"name\": \"wt\", \"value\": 2.465 }, {\"name\": \"qsec\", \"value\": 20.01 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 22
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 15.5 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 318 }, {\"name\": \"hp\", \"value\": 150 }, {\"name\": \"drat\", \"value\": 2.76 }, {\"name\": \"wt\", \"value\": 3.52 }, {\"name\": \"qsec\", \"value\": 16.87 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 23
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 15.2 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 304 }, {\"name\": \"hp\", \"value\": 150 }, {\"name\": \"drat\", \"value\": 3.15 }, {\"name\": \"wt\", \"value\": 3.435 }, {\"name\": \"qsec\", \"value\": 17.3 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 24
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 13.3 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 350 }, {\"name\": \"hp\", \"value\": 245 }, {\"name\": \"drat\", \"value\": 3.73 }, {\"name\": \"wt\", \"value\": 3.84 }, {\"name\": \"qsec\", \"value\": 15.41 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 25
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 19.2 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 400 }, {\"name\": \"hp\", \"value\": 175 }, {\"name\": \"drat\", \"value\": 3.08 }, {\"name\": \"wt\", \"value\": 3.845 }, {\"name\": \"qsec\", \"value\": 17.05 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 0 }, {\"name\": \"gear\", \"value\": 3 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 26
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 27.3 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 79 }, {\"name\": \"hp\", \"value\": 66 }, {\"name\": \"drat\", \"value\": 4.08 }, {\"name\": \"wt\", \"value\": 1.935 }, {\"name\": \"qsec\", \"value\": 18.9 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 1 }] }"
#> ------------------------------------------------------------ 
#> : 27
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 26 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 120.3 }, {\"name\": \"hp\", \"value\": 91 }, {\"name\": \"drat\", \"value\": 4.43 }, {\"name\": \"wt\", \"value\": 2.14 }, {\"name\": \"qsec\", \"value\": 16.7 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 5 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 28
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 30.4 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 95.1 }, {\"name\": \"hp\", \"value\": 113 }, {\"name\": \"drat\", \"value\": 3.77 }, {\"name\": \"wt\", \"value\": 1.513 }, {\"name\": \"qsec\", \"value\": 16.9 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 5 }, {\"name\": \"carb\", \"value\": 2 }] }"
#> ------------------------------------------------------------ 
#> : 29
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 15.8 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 351 }, {\"name\": \"hp\", \"value\": 264 }, {\"name\": \"drat\", \"value\": 4.22 }, {\"name\": \"wt\", \"value\": 3.17 }, {\"name\": \"qsec\", \"value\": 14.5 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 5 }, {\"name\": \"carb\", \"value\": 4 }] }"
#> ------------------------------------------------------------ 
#> : 30
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 19.7 }, {\"name\": \"cyl\", \"value\": 6 }, {\"name\": \"disp\", \"value\": 145 }, {\"name\": \"hp\", \"value\": 175 }, {\"name\": \"drat\", \"value\": 3.62 }, {\"name\": \"wt\", \"value\": 2.77 }, {\"name\": \"qsec\", \"value\": 15.5 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 5 }, {\"name\": \"carb\", \"value\": 6 }] }"
#> ------------------------------------------------------------ 
#> : 31
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 15 }, {\"name\": \"cyl\", \"value\": 8 }, {\"name\": \"disp\", \"value\": 301 }, {\"name\": \"hp\", \"value\": 335 }, {\"name\": \"drat\", \"value\": 3.54 }, {\"name\": \"wt\", \"value\": 3.57 }, {\"name\": \"qsec\", \"value\": 14.6 }, {\"name\": \"vs\", \"value\": 0 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 5 }, {\"name\": \"carb\", \"value\": 8 }] }"
#> ------------------------------------------------------------ 
#> : 32
#> [1] "{\"inputs\": [ {\"name\": \"mpg\", \"value\": 21.4 }, {\"name\": \"cyl\", \"value\": 4 }, {\"name\": \"disp\", \"value\": 121 }, {\"name\": \"hp\", \"value\": 109 }, {\"name\": \"drat\", \"value\": 4.11 }, {\"name\": \"wt\", \"value\": 2.78 }, {\"name\": \"qsec\", \"value\": 18.6 }, {\"name\": \"vs\", \"value\": 1 }, {\"name\": \"am\", \"value\": 1 }, {\"name\": \"gear\", \"value\": 4 }, {\"name\": \"carb\", \"value\": 2 }] }"